test: use Physician on EntityBuilder instead User
