+++ backend/src/test/java/br/ifsp/demo/presentation/controller/appointment/AppointmentControllerTest.java
@@ -1,8 +1,6 @@
 package br.ifsp.demo.presentation.controller.appointment;
 
-import br.ifsp.demo.presentation.security.user.User;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
+import br.ifsp.demo.presentation.BaseApiIntegrationTest;
 
 
 class AppointmentControllerTest extends BaseApiIntegrationTest {
